DEV: Remove FlagQuery class and old code
