+++ lib/flag_query.rb
@@ -220,15 +220,4 @@ module FlagQuery
       users: User.where(id: user_ids)
     }
   end
-
-  def self.excerpt(cooked)
-    excerpt = Post.excerpt(cooked, 200, keep_emoji_images: true)
-    # remove the first link if it's the first node
-    fragment = Nokogiri::HTML.fragment(excerpt)
-    if fragment.children.first == fragment.css("a:first").first && fragment.children.first
-      fragment.children.first.remove
-    end
-    fragment.to_html.strip
-  end
-
 end
